+++ src/main/java/com/strubium/openengie/OpenEngineering.java
@@ -55,12 +55,12 @@ public class OpenEngineering {
         GameRegistry.registerWorldGenerator(new OreGenerator(), 0);
 
 
-        GameRegistry.addSmelting(ModBlocks.ORE_ALUMINUM, new ItemStack(ModItems.INGOT_ALUMINUM), 0.7f);
+        GameRegistry.addSmelting(ModBlocks.ORE_ALUMINUM, new ItemStack(ModItems.INGOT_ALUMINUM), 0.4f);
         GameRegistry.addSmelting(ModBlocks.ORE_COPPER, new ItemStack(ModItems.INGOT_COPPER), 0.7f);
-        GameRegistry.addSmelting(ModBlocks.ORE_LEAD, new ItemStack(ModItems.INGOT_LEAD), 0.7f);
-        GameRegistry.addSmelting(ModBlocks.ORE_NICKEL, new ItemStack(ModItems.INGOT_NICKEL), 0.7f);
-        GameRegistry.addSmelting(ModBlocks.ORE_SILVER, new ItemStack(ModItems.INGOT_SILVER), 0.7f);
-        GameRegistry.addSmelting(ModBlocks.ORE_URANIUM, new ItemStack(ModItems.INGOT_URANIUM), 0.7f);
+        GameRegistry.addSmelting(ModBlocks.ORE_LEAD, new ItemStack(ModItems.INGOT_LEAD), 1.9f);
+        GameRegistry.addSmelting(ModBlocks.ORE_NICKEL, new ItemStack(ModItems.INGOT_NICKEL), 0.6f);
+        GameRegistry.addSmelting(ModBlocks.ORE_SILVER, new ItemStack(ModItems.INGOT_SILVER), 0.8f);
+        GameRegistry.addSmelting(ModBlocks.ORE_URANIUM, new ItemStack(ModItems.INGOT_URANIUM), 2.0f);
 
         Registry.registerOreDict();
     }